+++ module/data/fields/types.d.ts
@@ -1,0 +1,3 @@
+import { IdentifierMixin } from "./identifier.mjs";
+
+export interface IdentifierModel extends InstanceType<ReturnType<typeof IdentifierMixin>> {}