+++ module/data/item/quality.mjs
@@ -1,8 +1,15 @@
 import { CollectionField } from "@models/fields/object.mjs";
-import { CurseborneItemBase } from "./base.mjs";
 import { Complication, Enhancement } from "@models/modifiers.mjs";
+import { CurseborneItemBase } from "./base.mjs";
 
 export class AdversaryQuality extends CurseborneItemBase {
+	/** @inheritDoc */
+	static metadata = Object.freeze({
+		...super.metadata,
+		type: "quality",
+		identifierCollectionName: "qualities",
+	});
+
 	/** @inheritDoc */
 	static defineSchema() {
 		const fields = foundry.data.fields;